backprop

+++ src/Network/Csv.java
@@ -11,14 +11,14 @@ import java.util.List;
 import java.util.Random;
 
 public class Csv {
-    public static List<float[]> readDataFromCSV(String filePath) {
-        List<float[]> data = new ArrayList<>();
+    public static List<double[]> readDataFromCSV(String filePath) {
+        List<double[]> data = new ArrayList<>();
         try (CSVReader reader = new CSVReader(new FileReader(filePath))) {
-            float[] line;
+            double[] line;
             String[] temp;
             reader.readNext();
             while ((temp = reader.readNext()) != null) {
-                line = new float[temp.length];
+                line = new double[temp.length];
                 for (int i = 0; i < temp.length; i++) {
                     line[i] = Integer.parseInt(temp[i]);
                 }
@@ -30,27 +30,27 @@ public class Csv {
         return data;
     }
 
-    public static void shuffleArray(float[][] array) {
+    public static void shuffleArray(double[][] array) {
         Random rnd = new Random();
         for (int i = array.length - 1; i > 0; i--) {
             int index = rnd.nextInt(i + 1);
-            float[] temp = array[index];
+            double[] temp = array[index];
             array[index] = array[i];
             array[i] = temp;
         }
     }
 
-    public static float[] getColumn(float[][] array, int index) {
-        float[] column = new float[array.length];
+    public static double[] getColumn(double[][] array, int index) {
+        double[] column = new double[array.length];
         for (int i = 0; i < array.length; i++) {
             column[i] = array[i][index];
         }
         return column;
     }
 
-    public static float[][] removeColumn(float[][] array, int index) {
+    public static double[][] removeColumn(double[][] array, int index) {
         return Arrays.stream(array)
                 .map(row -> Arrays.copyOfRange(row, 1, row.length))
-                .toArray(float[][]::new);
+                .toArray(double[][]::new);
     }
 }
